deployment/LLM_bot/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
asyncio==3.4.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
diskcache==5.6.3
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.6
Jinja2==3.1.2
llama_cpp_python==0.2.27
logging==0.4.9.6
magic-filter==1.0.12
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
sympy==1.12
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
